requirements.txt
fastapi==0.115.0
requests==2.32.3
torch==2.4.1+cu121
transformers==4.48.0

python-dotenv~=1.0.1
nltk~=3.9.1